Refactor map view marker builders into dedicated part.
Split init_game_map view marker and capital helpers out of the cells/units part so each module stays focused and below review-size limits while preserving behavior.
